requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-api-response==0.1.0
django-redis==5.4.0
django-environ==0.9.0
django-ckeditor==6.3.2
django-cors-headers==4.6.0

beautifulsoup4==4.12.3
bleach==6.2.0

pyotp==2.9.0
qrcode==8.0
pillow==11.0.0

uvicorn==0.32.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2==2.9.4

djoser==2.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
djangorestframework-simplejwt==5.3.1

django-storages==1.14.4
boto3==1.35.15
botocore==1.35.15
cryptography==41.0.7
rsa==4.9

django-axes==7.0.0